Crossfire's emulator
This is an emulator for crossfire. Actually it's just a base, which you can just login in with an account, join a server and a channel, create a room. The source code still needs to be developed for years i think. But i think it could be useful for anyone who wants to realize how CF packets are going.